rebase: add todo-modifier commands
Teach rebase to accept options to redefine which todo-list commands to
perform for specific commits. Accept multiple instances of each option
to apply that command to multiple refs.

Add these options to the rebase command:
  --drop=ref
  --edit=ref
  --fixup=ref
  --pick=ref
  --reword=ref
  --squash=ref

For any refs given with these new options, modify the todo-list to replace
the "pick" command with the given command instead.  If no upstream is
given, choose a common parent to the given "command" option refs as
the upstream so at least the option refs are included in the todo-list.

WIP: Should die if all refs are not in the chosen todo-list
     Should die if any ref is mentioned for two different commands
     Any option with no ref could mean to use this option as the default
     Should explain need for --pick or delete it

Signed-off-by: Phil Hord <phord@purestorage.com>