WIP rebase: add switches for fixup and squash
Add two new edit-todo switches to let the user specify commits to
be squashed or fixed-up.  These switches are of dubious worth since
the ordering is unlikely to be correct without manual intervention,
but I like have the switches for completeness.

I am wondering if these should be changed to --fixup-onto and
--squash-onto , like this:

    # squash bar onto foo
    git rebase [--squash-onto foo [bar]]

But I think this would be very confusing to have a switch that
takes two arguments. Do they need to be joined somehow?

For that matter, maybe --pick-onto can be added as well to allow
the user to move a commit in the list.  But this may get confusing
when the user picks a commit onto another commit which itself was moved.

Currently this commit adds these use cases:
    '--fixup <ref>' to fixup a specific commit,
    '--squash <ref>' to squash a specific commit

Signed-off-by: Phil Hord <phil.hord@gmail.com>